Relocate `Proto3.Suite.DotProto.JSONPB` => `Proto3.Suite.JSONPB`
